refactor: ui, 도메인 분리
