+++ package.json
@@ -21,6 +21,7 @@
     "babel-jest": "^29.3.1",
     "cypress": "^12.6.0",
     "eslint": "^8.34.0",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-prettier": "^2.7.0",
     "jest": "^29.3.1",
     "prettier": "^2.8.4",
