+++ src/main.js
@@ -19,7 +19,7 @@ addEventListener('load', () => {
 
   main.appendChild(RestaurantList());
 
-  const { modal: restaurantAddModal, open } = RestaurantAddModal();
+  const { modal: restaurantAddModal, open, close } = RestaurantAddModal();
 
   main.appendChild(restaurantAddModal);
 
@@ -27,4 +27,30 @@ addEventListener('load', () => {
   plusButton.addEventListener('click', () => {
     open();
   });
+
+  $('.button--secondary').addEventListener('click', () => {
+    close();
+  });
+
+  const Form = $('form');
+  Form.addEventListener('submit', (e) => {
+    e.preventDefault();
+
+    const formData = new FormData(Form);
+    const data = Object.fromEntries(formData.entries()); // JSON 형태로 변환
+    handleSubmit(data);
+    Form.reset();
+    close();
+  });
 });
+
+const handleSubmit = ({ name, distance, description, category }) => {
+  const item = RestaurantItem({
+    name,
+    distance,
+    description,
+    icon: RestaurantIcon({ src: `../images/category-${category}.png`, alt: category }),
+  });
+
+  const restaurantList = $('.restaurant-list').appendChild(item);
+};
